requirements.txt
textract==1.6.3
mysql_connector_repackaged==0.3.1
mysql-connector==2.2.9